Post-writeup cleanup
Writing the paper has helped clarify some finer points, especially with
regards to calibration. It also establishes consistent notation.
These changes align the code with the paper’s notation and substantially
simplify the use of the Calibrate task:
- No need to use custom ModelColl class; any sized iterable will work.
- Previously the synthetic PPF curves for the A and B models were recreated
  for each data point. This is unnecessary and wasteful, but perhaps
  even worse, made calling `Calibrate` much more complicated, because
  we needed to specify callable models. Now we just provided the
  pre-computed PPF as a `interp1d` callable.
- To support this, the `make_empirical_ppf` utility was added. Having
  users call this themselves is good, because it makes it easy for them
  to visually check the PPF.
- We also make `scipy.interp1d` serializable, which should cover most
  use cases.

With the goal of putting this on PyPI, and since the `emd` package name
is already taken, we also rename the package to `emd_falsify`.

Still WIP (other than some renaming): viz.py